feat(admin-panel): add roles section, seeding roles, and restrict unauthorized access
- Added roles management section to the admin panel.
- Seeded roles data into the database.
- Restricted unauthorized users from accessing the admin panel.
- Improved 404 page layout to redirect unauthorized users to the homepage.

+++ Modules/PYB/Category/Http/Controllers/CategoryController.php
@@ -2,10 +2,11 @@
 
 namespace PYB\Category\Http\Controllers;
 
+use PYB\Category\Models\Category;
 use App\Http\Controllers\Controller;
-use PYB\Category\Http\Requests\CategoryRequest;
-use PYB\Category\Repositories\CategoryRepo;
 use PYB\Category\Services\CategoryService;
+use PYB\Category\Repositories\CategoryRepo;
+use PYB\Category\Http\Requests\CategoryRequest;
 
 class CategoryController extends Controller
 {
@@ -20,24 +21,30 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('index', Category::class);
+
+        return view('Panel::index');
         $categories = $this->repo->index()->paginate(10);
         return view('Category::index', compact('categories'));
     }
 
     public function create()
     {
+        $this->authorize('index', Category::class);
         $categories = $this->repo->index()->get();
         return view('Category::create', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('index', Category::class);
         $this->service->store($request);
         return to_route('categories.index') ->with(['success_store' => 'دسته بندی ساخته شد.'])  ;
     }
 
     public function edit($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $categories = $this->repo->index()->where('id', '!=', $category->id)->get();
         return view('Category::edit', compact('category'));
@@ -45,17 +52,20 @@ class CategoryController extends Controller
 
     public function update(CategoryRequest $request, $id)
     {
+        $this->authorize('index', Category::class);
         $this->service->update($request, $id);
         return to_route('categories.index') ->with(['success_update' => 'دسته بندی بروزرسانی شد.'])  ;
     }
 
     public function destroy($id)
     {
+        $this->authorize('index', Category::class);
         $this->repo->delete($id);
         return to_route('categories.index')->with(['success_delete' => 'دسته بندی حذف شد.']);
     }
     public function changeStatus($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $this->repo->changeStatus($category);
